Run experimentserver from gui async
