+++ ert/experiment_server/__init__.py
@@ -1,7 +1,5 @@
-from ._server import ExperimentServer
+from ._server import ExperimentServer, start_experiment_server
 from ._state_machine import StateMachine
+from ._experiment_protocol import Experiment
 
-__all__ = (
-    "ExperimentServer",
-    "StateMachine",
-)
+__all__ = ("ExperimentServer", "StateMachine", "Experiment", "start_experiment_server")
